Fix Reflex Var comparison error in bucket funding bar
row["budget"] is an ObjectItemOperation Var inside rx.foreach — it
doesn't support Python's native > operator. Use .to(float) to cast it
to a typed Var before comparing with 0.

+++ reflex_app/cura/components/buckets.py
@@ -652,7 +652,7 @@ def _bucket_row(row: dict) -> rx.Component:
 
             # Funding bar: alloc / budget (ZBB primary metric)
             rx.cond(
-                row["budget"] > 0,
+                row["budget"].to(float) > 0,
                 rx.box(
                     rx.box(
                         class_name="prog-fill",
